+++ src/main/java/pigcart/cosycritters/mixin/access/ParticleEngineAccessor.java
@@ -1,16 +1,23 @@
 package pigcart.cosycritters.mixin.access;
 
 import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
+import net.minecraft.client.particle.Particle;
 import net.minecraft.client.particle.ParticleEngine;
 //? if >= 1.21.9 {
 /*import net.minecraft.core.particles.ParticleLimit;
 *///?} else {
+import net.minecraft.client.particle.ParticleRenderType;
 import net.minecraft.core.particles.ParticleGroup;
 import net.minecraft.client.renderer.texture.TextureAtlas;
 //?}
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 import org.spongepowered.asm.mixin.gen.Invoker;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+
+import java.util.Map;
+import java.util.Queue;
 
 @Mixin(ParticleEngine.class)
 public interface ParticleEngineAccessor {
@@ -19,24 +26,6 @@ public interface ParticleEngineAccessor {
     @Accessor
     TextureAtlas getTextureAtlas();
     //?}
-
-    @Accessor
-    Object2IntOpenHashMap<
-            //? if >= 1.21.9 {
-            /*ParticleLimit
-            *///?} else {
-            ParticleGroup
-            //?}
-            > getTrackedParticleCounts();
-
-    @Invoker
-    boolean callHasSpaceInParticleLimit(
-            //? if >= 1.21.9 {
-            /*ParticleLimit
-            *///?} else {
-            ParticleGroup
-            //?}
-            group);
 
     @Invoker
     void callClearParticles();